+++ swagger_parser/lib/src/generator/models/universal_data_class.dart
@@ -1,7 +1,13 @@
 /// Universal template for containing information about component
 abstract class UniversalDataClass {
-  const UniversalDataClass({required this.name});
+  const UniversalDataClass({
+    required this.name,
+    this.description,
+  });
 
   /// Name of the class
   final String name;
+
+  /// Description of the class
+  final String? description;
 }